pyproject.toml
[tool.poetry]
name = "Phone Pages"
version = "1.0"
description = "The Directory is an API application(DRF) for searching phone numbers and information about organizations"
authors = ["Konstantin Raikhert <konstantinraihert@yandex.ru>"]
license = "MIT"


[tool.black]
line-length = 79
include = '\.pyi?$'
extend-exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | project-settings/settings
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.poetry.dependencies]
python = "^3.8"
django = "--dev"
djangorestframework = "^3.12.4"
django-extensions = "^3.1.3"
django-filter = "^2.4.0"
drf-yasg2 = "^1.19.4"
black = "^21.8b0"
flake8 = "^3.9.2"
isort = "^5.9.3"
psycopg2-binary = "^2.9.1"
python-environ = "^0.4.54"
factory-boy = "^3.2.0"
django-phone-field = "^1.8.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
